+++ examples/scripts/compare_lead_acid.py
@@ -17,9 +17,9 @@ else:
 
 # load models
 models = [
-    pybamm.lead_acid.LOQS(),
-    # pybamm.lead_acid.FOQS(),
-    pybamm.lead_acid.Composite(),
+    # pybamm.lead_acid.LOQS(),
+    # # pybamm.lead_acid.FOQS(),
+    # pybamm.lead_acid.Composite(),
     pybamm.lead_acid.Full(),
 ]